rename Huffman*Encoding to CanonicalHuffman*Encoding
